create tables - accounts and types

+++ database/migrations/2019_07_30_134733_create_accounts_table.php
@@ -15,9 +15,9 @@ class CreateAccountsTable extends Migration
     {
         Schema::create('accounts', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('account_type_id')->unsigned();
+            $table->bigInteger('account_type_id')->unsigned();
             $table->foreign('account_type_id')->references('id')->on('account_types')->onDelete('cascade');
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('account_number')->unique();
             $table->timestamps();
